refactor: // todo
feat:
support display inode of file/dir
