+++ osbased/inode_linux.go
@@ -1,0 +1,17 @@
+//go:build linux
+
+package osbased
+
+import (
+	"os"
+	"strconv"
+	"syscall"
+)
+
+func Inode(info os.FileInfo) string {
+	stat, ok := info.Sys().(*syscall.Stat_t)
+	if ok {
+		return strconv.FormatUint(stat.Ino, 10)
+	}
+	return ""
+}
