+++ osbased/inode_windows.go
@@ -1,0 +1,7 @@
+package osbased
+
+import "os"
+
+func Inode(info os.FileInfo) string {
+	return "-"
+}
